// src/models/billing.model.ts
export interface IBilling {
    name: string;
    certNo: string;
    contractNo: string;
    contractType: number;
    address: string;
    feeStatus: string;
    terms: string;
    money: number;
    payedDate: string;
    status: number;
    communityId: string;
    invoiceNo: string;
    invoiceUser: string;
    remark: string;
}